move defaultCustomization para consts
