+++ src/features/acf/frontend/common/WithCustomPrint/consts.ts
@@ -1,0 +1,6 @@
+export const defaultCustomization = {
+    grouping: true,
+    splitGroupPerPage: false,
+    order: false,
+    orderGroup: (): number => 0,
+};
